requirements.txt
django==5.2.4
psycopg==3.2.9
pillow==11.3.0
djangorestframework==3.16.0